Move new files out of systems/rigid_body_plant into multibody
This includes changing include file references
