+++ drake/multibody/rigid_body_plant/contact_detail.h
@@ -4,7 +4,7 @@
 
 #include "drake/common/drake_export.h"
 #include "drake/common/eigen_types.h"
-#include "drake/systems/plants/rigid_body_plant/contact_force.h"
+#include "drake/multibody/rigid_body_plant/contact_force.h"
 
 namespace drake {
 namespace systems {
