+++ drake/multibody/rigid_body_plant/contact_resultant_force_calculator.cc
@@ -1,5 +1,5 @@
 #include "drake/multibody/rigid_body_plant/contact_resultant_force_calculator.h"
-#include "drake/systems/plants/rigid_body_plant/point_contact_detail.h"
+#include "drake/multibody/rigid_body_plant/point_contact_detail.h"
 
 namespace drake {
 namespace systems {
