+++ drake/multibody/rigid_body_plant/contact_resultant_force_calculator.h
@@ -5,7 +5,7 @@
 #include "drake/common/drake_export.h"
 #include "drake/common/eigen_types.h"
 #include "drake/multibody/rigid_body_plant/contact_force.h"
-#include "drake/systems/plants/rigid_body_plant/contact_detail.h"
+#include "drake/multibody/rigid_body_plant/contact_detail.h"
 
 namespace drake {
 namespace systems {
