+++ drake/multibody/rigid_body_plant/contact_results.h
@@ -3,8 +3,8 @@
 #include <vector>
 
 #include "drake/common/drake_export.h"
-#include "drake/systems/plants/collision/Element.h"
-#include "drake/systems/plants/rigid_body_plant/contact_info.h"
+#include "drake/multibody/collision/Element.h"
+#include "drake/multibody/rigid_body_plant/contact_info.h"
 
 namespace drake {
 namespace systems {
