+++ drake/multibody/rigid_body_plant/point_contact_detail.cc
@@ -1,4 +1,4 @@
-#include "drake/systems/plants/rigid_body_plant/point_contact_detail.h"
+#include "drake/multibody/rigid_body_plant/point_contact_detail.h"
 
 namespace drake {
 namespace systems {
